backend/target_reader/pyproject.toml
[project]
name = "target-reader"
version = "0.1.0"
description = "arch-stats: Target Reader Module"
authors = [
    { name = "Juan Pablo Molina Matute", email = "jpmolinamatute@gmail.com" }
]
requires-python = ">=3.12.7"
dependencies = []

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pylint>=3.3.7",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]
